refactor: move quiz upgrader to live with quiz schemas
- Move quizUpgrader from features/upgrades to protocol/schemas/quiz/conversion
- Update all import paths in consuming files
- Fix TypeScript and linting issues in test file
- Keep quiz upgrade logic co-located with other quiz conversion utilities
